refactor: Avoid importing server code inside client code
