+++ app/graphql/GraphqlProvider.tsx
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+import { Provider } from "urql";
+
+import { client } from "./client";
+
+export const GraphqlProvider = ({ children }: { children: ReactNode }) => {
+  return <Provider value={client}>{children}</Provider>;
+};
